+++ changelog.md
@@ -9,6 +9,7 @@
 - customizable chatters
 - chat on by default
 - customizable chat pos
+- viewer count thats customizable
 ## v1.0.1
 - Removed debug logs
 - Fixed readme